#9 - Moving the tests into a separate folder
