+++ src/handlers/__tests__/complianceReportPdfHandler.test.js
@@ -2,7 +2,7 @@ const express = require('express');
 const supertest = require('supertest');
 const fileUpload = require('express-fileupload');
 const path = require('path');
-const { complianceReportPdfHandler } = require('./complianceReportPdfHandler');
+const { complianceReportPdfHandler } = require('../complianceReportPdfHandler');
 
 const URL = '/api/compliance-report';
 const app = express();
@@ -58,7 +58,7 @@ describe('Handling an English file', () => {
   ];
 
   describe('Sending the english file as part of multipart formdata', () => {
-    const SAMPLE_PDF = path.join(__dirname, 'compliance_report__en.pdf');
+    const SAMPLE_PDF = path.join(__dirname, 'compliance-report-english.pdf');
 
     let response;
 
@@ -84,7 +84,7 @@ describe('Handling an English file', () => {
 });
 
 //   describe('Sending the file as a BASE64 encoded string', () => {
-//     const SAMPLE_PDF = path.join(__dirname, 'report-en.pdf');
+//     const SAMPLE_PDF = path.join(__dirname, 'diagnostic-report-english.pdf');
 //     const base64 = fs.readFileSync(SAMPLE_PDF, { encoding: 'base64' });
 //
 //     let response;
@@ -224,7 +224,7 @@ describe('Handling an English file', () => {
 //   ];
 //
 //   describe('Sending the file as part of multipart formdata', () => {
-//     const SAMPLE_PDF = path.join(__dirname, 'report-portugese.pdf');
+//     const SAMPLE_PDF = path.join(__dirname, 'diagnostic-report-portugese.pdf');
 //
 //     let response;
 //
@@ -251,7 +251,7 @@ describe('Handling an English file', () => {
 //
 // describe('Gender should be translated to the english word', () => {
 //   it('should translate portugese', async () => {
-//     const SAMPLE_PDF = path.join(__dirname, 'report-portugese.pdf');
+//     const SAMPLE_PDF = path.join(__dirname, 'diagnostic-report-portugese.pdf');
 //     const response = await supertest(app)
 //       .post(URL)
 //       .attach('pdf', SAMPLE_PDF);
