+++ src/handlers/__tests__/diagnostic-report-portugese.pdf
